[IMP] web: convert core/domain_selector_dialog tests to hoot
task-id: 3705027
